Split functions into modules
